// src/api/openapi-client/models/CooperationCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Retirement‐Matching API
 * REST endpoints for user auth (via magic‐link), retirees, career elements, reference data, startups, job postings, matching, conversations, and hire invites. Uses a JWT set in a cookie. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CooperationCreate
 */
export interface CooperationCreate {
    /**
     * 
     * @type {string}
     * @memberof CooperationCreate
     */
    retireeId: string;
    /**
     * 
     * @type {string}
     * @memberof CooperationCreate
     */
    jobPostingId: string;
}

/**
 * Check if a given object implements the CooperationCreate interface.
 */
export function instanceOfCooperationCreate(value: object): value is CooperationCreate {
    if (!('retireeId' in value) || value['retireeId'] === undefined) return false;
    if (!('jobPostingId' in value) || value['jobPostingId'] === undefined) return false;
    return true;
}

export function CooperationCreateFromJSON(json: any): CooperationCreate {
    return CooperationCreateFromJSONTyped(json, false);
}

export function CooperationCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): CooperationCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'retireeId': json['retireeId'],
        'jobPostingId': json['jobPostingId'],
    };
}

export function CooperationCreateToJSON(json: any): CooperationCreate {
    return CooperationCreateToJSONTyped(json, false);
}

export function CooperationCreateToJSONTyped(value?: CooperationCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'retireeId': value['retireeId'],
        'jobPostingId': value['jobPostingId'],
    };
}