// src/api/openapi-client/models/CooperationUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Retirement‐Matching API
 * REST endpoints for user auth (via magic‐link), retirees, career elements, reference data, startups, job postings, matching, conversations, and hire invites. Uses a JWT set in a cookie. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CooperationUpdate
 */
export interface CooperationUpdate {
    /**
     * 
     * @type {string}
     * @memberof CooperationUpdate
     */
    status?: CooperationUpdateStatusEnum;
}


/**
 * @export
 */
export const CooperationUpdateStatusEnum = {
    Pending: 'pending',
    Accepted: 'accepted',
    Declined: 'declined'
} as const;
export type CooperationUpdateStatusEnum = typeof CooperationUpdateStatusEnum[keyof typeof CooperationUpdateStatusEnum];


/**
 * Check if a given object implements the CooperationUpdate interface.
 */
export function instanceOfCooperationUpdate(value: object): value is CooperationUpdate {
    return true;
}

export function CooperationUpdateFromJSON(json: any): CooperationUpdate {
    return CooperationUpdateFromJSONTyped(json, false);
}

export function CooperationUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): CooperationUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function CooperationUpdateToJSON(json: any): CooperationUpdate {
    return CooperationUpdateToJSONTyped(json, false);
}

export function CooperationUpdateToJSONTyped(value?: CooperationUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
    };
}